Brute Force V.1.0
New bird detection algorithm which uses logic to eliminate as much area as possible, then calculates TDOA for an emitter at any given location inside the realm of possible area.  If the theoretical TDOA matches the real (to within 1%), we will consider it.  This generates a thick "fan", or "flat" cone with low aperture where each point represents a possible location of the bird.  Next step will be to generate an equation for line of fit, in order to more accurately predict the real location.  Then, I will make the code more universal.  I am currently assuming some starting condition for the bird's location, so right now the code only considers quadrant 1.